requirements.txt
attrs==22.1.0
bytecode==0.15.0
cattrs==22.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==7.1.2
cmake==3.27.2
ddsketch==2.0.4
ddtrace==2.11.1
Deprecated==1.2.14
envier==0.5.2
exceptiongroup==1.0.4
Flask==1.1.2
forbiddenfruit==0.1.4
googleapis-common-protos==1.66.0
grpcio==1.69.0
gunicorn==22.0.0
idna==3.10
importlib-metadata==6.0.1
intervaltree==3.0.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==4.17.1
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-enum==1.5.1
msgpack==1.0.0
mypy-extensions==0.4.3
opentelemetry-api==1.29.0
opentelemetry-distro==0.50b0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==21.3
protobuf==4.21.6
pycparser==2.22
pyparsing==3.0.9
pyroscope-io==0.8.8
pyroscope-otel==0.4.0
pyrsistent==0.19.2
requests==2.32.3
six==1.16.0
sortedcontainers==2.1.0
stringcase==1.2.0
tenacity==8.1.0
typing-extensions==3.7.4.2
typing-inspect==0.6.0
urllib3==2.3.0
Werkzeug==1.0.1
wrapt==1.15.0
xmltodict==0.13.0
zipp==3.19.1
